Week5 Done

+++ SERVICES/STD_TYPES.h
@@ -8,8 +8,8 @@ typedef signed long int    s32;
 
 /* Unsigned Types */
 typedef unsigned char        u8;
-typedef unsigned short int   u16;
-typedef unsigned long int    u32;
+typedef unsigned int   u16;
+typedef unsigned long    u32;
 
 /* Floating Types */
 typedef float       f32;
